IMPALA-8363: fix e2e start with impala_log_dir
This commit fixes the `CustomClusterTestSuite` to wait for the
correct number of executors when `impala_log_dir` is specified
in the test decorator.  Previously, the default value of 3
was always used, regardless of `cluster_size`.

Testing:

- Ran tests/authorization/test_ranger.py with custom
`impala_log_dir` and `cluster_size` arguments and
observed pass.  Observed to fail before changes

- Ran all original e2e tests

Change-Id: I4f46f40474b4b380abe88647a37e8e4d2231d745